+++ vnext/Microsoft.ReactNative.Cxx/ReactContext.h
@@ -7,6 +7,7 @@
 
 #include <string_view>
 #include "JSValueWriter.h"
+#include "ReactDispatcher.h"
 #include "ReactPropertyBag.h"
 
 namespace winrt::Microsoft::ReactNative {
@@ -24,11 +25,15 @@ struct ReactContext {
   }
 
   explicit operator bool() const noexcept {
-    return static_cast<bool>(m_handle);
+    return m_handle ? true : false;
   }
 
   ReactPropertyBag Properties() const noexcept {
     return ReactPropertyBag{m_handle.Properties()};
+  }
+
+  ReactDispatcher UIDispatcher() const noexcept {
+    return ReactDispatcher{m_handle.UIDispatcher()};
   }
 
   // Call methodName JS function of module with moduleName.