+++ vnext/Microsoft.ReactNative/IReactDispatcher.h
@@ -3,23 +3,24 @@
 
 #pragma once
 #include "ReactDispatcherHelper.g.h"
-#include <ReactPropertyBag.h>
 #include <dispatchQueue/dispatchQueue.h>
 #include <winrt/Microsoft.ReactNative.h>
 
-namespace winrt::Microsoft::ReactNative {
+namespace winrt::Microsoft::ReactNative::implementation {
 
-struct ReactDispatcher : implements<ReactDispatcher, IReactDispatcher> {
+struct ReactDispatcher : implements<ReactDispatcher, winrt::default_interface<IReactDispatcher>> {
   ReactDispatcher() = default;
   ReactDispatcher(Mso::DispatchQueue &&queue) noexcept;
 
   bool HasThreadAccess() noexcept;
   void Post(ReactDispatcherCallback const &callback) noexcept;
 
+  static IReactDispatcher CreateSerialDispatcher() noexcept;
+
   static Mso::DispatchQueue GetUIDispatchQueue(IReactPropertyBag const &properties) noexcept;
 
   static IReactDispatcher UIThreadDispatcher() noexcept;
-  static ReactPropertyId<IReactDispatcher> UIDispatcherProperty() noexcept;
+  static IReactPropertyName UIDispatcherProperty() noexcept;
   static IReactDispatcher GetUIDispatcher(IReactPropertyBag const &properties) noexcept;
   static void SetUIThreadDispatcher(IReactPropertyBag const &properties) noexcept;
 
@@ -27,19 +28,19 @@ struct ReactDispatcher : implements<ReactDispatcher, IReactDispatcher> {
   Mso::DispatchQueue m_queue;
 };
 
-} // namespace winrt::Microsoft::ReactNative
-
-namespace winrt::Microsoft::ReactNative::implementation {
-
 struct ReactDispatcherHelper {
   ReactDispatcherHelper() = default;
+
+  static IReactDispatcher CreateSerialDispatcher() noexcept {
+    return ReactDispatcher::CreateSerialDispatcher();
+  }
 
   static IReactDispatcher UIThreadDispatcher() noexcept {
     return ReactDispatcher::UIThreadDispatcher();
   }
 
   static IReactPropertyName UIDispatcherProperty() noexcept {
-    return ReactDispatcher::UIDispatcherProperty().Handle();
+    return ReactDispatcher::UIDispatcherProperty();
   }
 };